+++ app/controllers/marketing_controller.ts
@@ -454,6 +454,8 @@ export default class MarketingController {
           ctaHref: i18n.t('marketing.home.demo.cta_href'),
           secondaryLabel: i18n.t('marketing.home.demo.secondary_label'),
           noCommitment: i18n.t('marketing.home.demo.no_commitment'),
+          tryDemoLabel: i18n.t('marketing.home.demo.try_demo_label'),
+          tryDemoSubtitle: i18n.t('marketing.home.demo.try_demo_subtitle'),
         },
         finalCta: {
           title: i18n.t('marketing.home.final_cta.title'),